tidy up display messages + moved processor into own file
